--- test/unit/generator/generate-statement.test.js
@@ -1,10 +1,0 @@
-
-describe('generate statement', () => {
-  afterEach(() => {
-    jest.clearAllMocks()
-  })
-
-  test('placeholder', async () => {
-    expect(true).toBeTruthy()
-  })
-})
